Restructure example-setting to be nested entities
Set up the example-setting so that parent entities include a
collection of the child entities. For example the "outwilds" territory
has a child entity of the "goblin" race. This will help in creating a
setting's category lists to choose from. As part of these changes, I
have a temporary hard coded schema added to the conn so that the
territories and races recursively expand when pulled.